chore: Update imports for editor move
