+++ src/ops/cargo_add/crate_spec.rs
@@ -4,7 +4,7 @@ use anyhow::Context as _;
 
 use super::Dependency;
 use crate::CargoResult;
-use crate::util::toml_mut::dependency::RegistrySource;
+use crate::workspace::editor::dependency::RegistrySource;
 use cargo_util_schemas::manifest::PackageName;
 
 /// User-specified crate
