+++ src/ops/cargo_add/mod.rs
@@ -28,15 +28,6 @@ use crate::util::OptVersionReq;
 use crate::util::cache_lock::CacheLockMode;
 use crate::util::edit_distance;
 use crate::util::style;
-use crate::util::toml_mut::dependency::Dependency;
-use crate::util::toml_mut::dependency::GitSource;
-use crate::util::toml_mut::dependency::MaybeWorkspace;
-use crate::util::toml_mut::dependency::PathSource;
-use crate::util::toml_mut::dependency::RegistrySource;
-use crate::util::toml_mut::dependency::Source;
-use crate::util::toml_mut::dependency::WorkspaceSource;
-use crate::util::toml_mut::manifest::DepTable;
-use crate::util::toml_mut::manifest::LocalManifest;
 use crate::workspace::Feature;
 use crate::workspace::FeatureValue;
 use crate::workspace::Features;
@@ -46,6 +37,15 @@ use crate::workspace::Registry;
 use crate::workspace::Summary;
 use crate::workspace::Workspace;
 use crate::workspace::dependency::DepKind;
+use crate::workspace::editor::dependency::Dependency;
+use crate::workspace::editor::dependency::GitSource;
+use crate::workspace::editor::dependency::MaybeWorkspace;
+use crate::workspace::editor::dependency::PathSource;
+use crate::workspace::editor::dependency::RegistrySource;
+use crate::workspace::editor::dependency::Source;
+use crate::workspace::editor::dependency::WorkspaceSource;
+use crate::workspace::editor::manifest::DepTable;
+use crate::workspace::editor::manifest::LocalManifest;
 use crate::workspace::parser::lookup_path_base;
 use crate::workspace::registry::PackageRegistry;
 use crate_spec::CrateSpec;
