+++ src/ops/cargo_remove.rs
@@ -2,10 +2,10 @@
 
 use crate::CargoResult;
 use crate::GlobalContext;
-use crate::util::toml_mut::manifest::DepTable;
-use crate::util::toml_mut::manifest::LocalManifest;
-use crate::util::toml_mut::manifest::MissingDependencyError;
 use crate::workspace::Package;
+use crate::workspace::editor::manifest::DepTable;
+use crate::workspace::editor::manifest::LocalManifest;
+use crate::workspace::editor::manifest::MissingDependencyError;
 
 /// Remove a dependency from a Cargo.toml manifest file.
 #[derive(Debug)]
